+++ Sources/HFAPI/Hub/Space.swift
@@ -54,7 +54,7 @@ public struct Space: Identifiable, Codable, Sendable {
     public let subdomain: String?
 
     /// The trending score.
-    public let trendingScore: Int?
+    public let trendingScore: Double?
 
     /// The used storage in bytes.
     public let usedStorage: Int?
